requirements.txt
# Core dependencies
paddleocr>=2.7
paddlepaddle>=2.5
faster-whisper>=0.10
moviepy>=1.0.3
opencv-python>=4.8
pyyaml>=6.0
typer>=0.9
zhipuai>=2.0

# Additional dependencies
numpy>=1.24
Pillow>=10.0
requests>=2.31
